Check for mode flags, set Emulation Mode as default if no flags supplied.
At least one of the define's `DEBUG_MODE` and `EMULATION_MODE` must be set for the library compilation
This is a safeguard for giving an understandable warning when neither are set, and for setting a default.
Not doing this will otherwise result in strange compilation errors.

While the makefile adequately handles these define's, the code might change later on in which the given situation occurs.

As an extra, this commit adds a `.gitignore` to the project.