Make debug output dynamicA
This centralizes all library code between `#ifdef DEBUG` blocks.

A new class `Debug` is added to the Library. This allows for debug output enabling in the code, without having to recompile the library.

The debug output can be turned on and off with a call to `Debug::enable()`. It is also possible to redirect debug output to a logfile.

In particular, it gets rid of `#ifdef DEBUG` statements in the library code. The way is now open to remove `DEBUG=1` as a make option,
which currently doesn't do anything any more.

Code has been added to example program `Tri` to show usage.