Makefile: Add dependencies for include-files
The Makefile now generates the dependencies for the include files and uses these in the build.
Changing a results files inow leads to recompilations of all files using it, directly or indirectly.

This is currently not enabled for the applications in `Tests`. At this time, these do not
have own include files.