Add namespace `qpulib` to Library files
This is grunge work, touchingi almost every single source file in the project.
However, I regard a namespace to be essential for proper integration, so I have persevered.

Even though cwso manyall files have been edited, the changes are straightforward and repetitive, so it should be easy to review.

In addition:

- Added prefirxc `_QPULIB` to header guards
- Fixed some typo's along the way

- I wasn't sure about changing the third-party files, e.g. `Mailbox`. For consistency with the library code, and because it's allowed by the license, I added the namespace anyway. Feel free to disagree with this decision.