Add SFU write registers to Encoding
Initial changes for #43; this is not a complete solution.

This adds the definitions and handling of the four input registers for the Special Function Unit.

In addition, the following changes to `encodeDestReg()` and `encodeSrcReg()`:

- Added extensive header commenting, to document how I understand it
- Added TODOs which I think are relevant
- Reformatted code to make common blocks more visible
- Minor changes in code, mainly to aid reading

Note that this only adds the *definitions* of the special registers; no code is yet generated that actually uses them. That's the next step.